+++ skgstat/__init__.py
@@ -5,6 +5,6 @@ from .Kriging import OrdinaryKriging
 from . import interfaces
 
 # set some stuff
-__version__ = '0.3.7'
+__version__ = '0.3.8'
 __author__ = 'Mirko Maelicke <mirko.maelicke@kit.edu>'
 __backend__ = 'matplotlib'